test(hps): refactor tests
